+++ NEWS.md
@@ -1,4 +1,11 @@
 
+### 1.7.1
+
+* Added Dirichlet-multinomial distribution
+* Code clean-up for categorical distribution
+* Improvements in documentation and examples (e.g. bivariate Poisson,
+  bivariate normal)
+
 ### 1.7.0
 
 * Cleaning namespace - now mathematical functions are explicitely